Recording: Settings handling fixes / overhaul. This adds the ability to handle per-screen settings, which can be different from the first screen (screen 0). Also fixed a couple of bugs regarding snapshot handling and persistence (committing, rolling back, ++) in that area. FE/VBoxManage now can also list the per-screen settings. Added some further @todos [build fix]. bugref:9286
svn:sync-xref-src-repo-rev: r152285

+++ src/VBox/Main/xml/Settings.cpp
@@ -1,4 +1,4 @@
-/* $Id: Settings.cpp 95639 2022-07-14 08:30:45Z andreas.loeffler@oracle.com $ */
+/* $Id: Settings.cpp 95641 2022-07-14 08:57:21Z andreas.loeffler@oracle.com $ */
 /** @file
  * Settings File Manipulation API.
  *
@@ -8007,7 +8007,7 @@ void MachineConfigFile::buildRecordingXML(xml::ElementNode &elmParent, const Rec
             pelmRecording->setAttribute("enabled", recording.common.fEnabled);
 
         /* Only serialize screens which have non-default settings. */
-        size_t cScreensToWrite = 0;
+        uint32_t cScreensToWrite = 0;
 
         RecordingScreenSettingsMap::const_iterator itScreen = recording.mapScreens.begin();
         while (itScreen != recording.mapScreens.end())
